refactor(desktop): isolate git workspace service
